# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Catalog moderation methods for skipper cms",
        "title": "Catalog service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/catalog": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменение имени элемента каталога",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputForms.ChangeCatalogNameInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Catalog"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавление элементов каталога",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputForms.AddCatalogInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Catalog"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление элемента каталога",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id элемента каталога",
                        "name": "catalog_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень элемента католога",
                        "name": "catalog_level",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outputForms.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outputForms.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outputForms.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/catalog/child": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение детей элемента каталога",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id родителя элемента",
                        "name": "parent_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень элемента католога",
                        "name": "catalog_level",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Catalog"
                            }
                        }
                    }
                }
            }
        },
        "/catalog/main-catalog": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение верхнего уровня католога",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MainCatalog"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "inputForms.AddCatalogInput": {
            "type": "object",
            "required": [
                "catalog_name"
            ],
            "properties": {
                "catalog_level": {
                    "type": "integer"
                },
                "catalog_name": {
                    "type": "string"
                },
                "catalog_parent_id": {
                    "type": "integer"
                }
            }
        },
        "inputForms.ChangeCatalogNameInput": {
            "type": "object",
            "properties": {
                "catalog_id": {
                    "type": "integer"
                },
                "catalog_level": {
                    "type": "integer"
                },
                "newCatalogName": {
                    "type": "string"
                }
            }
        },
        "models.Catalog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
        "models.MainCatalog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "outputForms.DeleteResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "outputForms.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}